Rename util package to utils to avoid name conflict with java.util.
